requirements.txt
annotated-types==0.7.0
anyio==4.8.0
autopep8==2.3.2
browser-cookie3==0.19.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
docopt==0.6.2
genshinstats==1.4.11.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lz4==4.3.3
openai==1.60.1
pipreqs==0.4.13
pycodestyle==2.12.1
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
yarg==0.1.10
